organize frontend webscket handlers
