+++ packages/client/src/components/chat/chat-project-sidebar.tsx
@@ -15,9 +15,9 @@ import { SelectedFilesList } from '@/components/projects/selected-files-list'
 import { SlidingSidebar } from '@/components/sliding-sidebar'
 import { PromptsList } from '../projects/prompts-list'
 import { useCopyClipboard } from '@/hooks/utility-hooks/use-copy-clipboard'
-import { useUpdateChatLinkSettings, useUnlinkChatTab } from '@/components/global-state/global-helper-hooks'
+import { useUpdateChatLinkSettings, useUnlinkChatTab } from '@/websocket-state/hooks/updaters/websocket-updater-hooks'
 import { useSelectedFiles, } from '@/hooks/utility-hooks/use-selected-files'
-import { useActiveChatTab, useProjectTab } from '../global-state/websocket-selector-hoooks'
+import { useActiveChatTab, useProjectTab } from '@/websocket-state/hooks/selectors/websocket-selector-hoooks'
 
 type ChatProjectSidebarProps = {
     linkedProjectTabId: string
