+++ packages/client/src/components/navigation/app-navbar.tsx
@@ -13,8 +13,8 @@ import { useQuery } from "@tanstack/react-query"
 import { useApi } from "@/hooks/use-api"
 import { HelpDialog } from "@/components/navigation/help-dialog"
 import { SettingsDialog } from "@/components/settings/settings-dialog"
-import { useUpdateActiveProjectTab } from "@/components/global-state/global-helper-hooks"
-import { useActiveProjectTab, useSettings } from "../global-state/websocket-selector-hoooks"
+import { useUpdateActiveProjectTab } from "@/websocket-state/hooks/updaters/websocket-updater-hooks"
+import { useActiveProjectTab, useSettings } from "@/websocket-state/hooks/selectors/websocket-selector-hoooks"
 
 export function AppNavbar() {
     const [openDialog, setOpenDialog] = useState(false)
