+++ packages/client/src/components/projects/file-tree/file-tree.tsx
@@ -27,8 +27,8 @@ import { toast } from 'sonner'
 import { buildNodeContent } from '@/components/projects/utils/projects-utils'
 import clsx from 'clsx'
 import { EditorType } from 'shared/src/global-state/global-state-schema'
-import { useGlobalStateCore } from '@/components/global-state/global-helper-hooks'
-import { useActiveProjectTab } from '@/components/global-state/websocket-selector-hoooks'
+import { useGlobalStateCore } from '@/websocket-state/hooks/updaters/websocket-updater-hooks'
+import { useActiveProjectTab } from '@/websocket-state/hooks/selectors/websocket-selector-hoooks'
 
 type SetSelectedFilesFunction = (updater: (prev: string[]) => string[]) => void;
