+++ packages/client/src/components/projects/project-settings-dialog.tsx
@@ -19,9 +19,9 @@ import {
 import { Input } from '@/components/ui/input'
 import { Slider } from '@/components/ui/slider'
 import { EDITOR_OPTIONS, type EditorType } from 'shared/src/global-state/global-state-schema'
-import { useUpdateActiveProjectTab, useUpdateGlobalStateKey } from '@/components/global-state/global-helper-hooks'
+import { useUpdateActiveProjectTab, useUpdateGlobalStateKey } from '@/websocket-state/hooks/updaters/websocket-updater-hooks'
 import { useSyncProjectInterval } from '@/hooks/api/use-projects-api'
-import { useActiveProjectTab, useSettings } from '../global-state/websocket-selector-hoooks'
+import { useActiveProjectTab, useSettings } from '@/websocket-state/hooks/selectors/websocket-selector-hoooks'
 
 
 export function ProjectSettingsDialog() {
