+++ packages/client/src/components/projects/selected-files-list.tsx
@@ -10,11 +10,11 @@ import { Input } from "../ui/input"
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogFooter } from "../ui/dialog"
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger, DropdownMenuSeparator, DropdownMenuSub, DropdownMenuSubTrigger, DropdownMenuSubContent, DropdownMenuRadioGroup, DropdownMenuRadioItem, DropdownMenuShortcut } from "../ui/dropdown-menu"
 import { toast } from "sonner"
-import { useGlobalStateCore, useUpdateProjectTabState } from "@/components/global-state/global-helper-hooks"
+import { useGlobalStateCore, useUpdateProjectTabState } from "@/websocket-state/hooks/updaters/websocket-updater-hooks"
 import { DotsHorizontalIcon } from "@radix-ui/react-icons"
 import { formatShortcut } from "@/lib/shortcuts"
 import { type UseSelectedFileReturn } from '@/hooks/utility-hooks/use-selected-files'
-import { useProjectTab } from "../global-state/websocket-selector-hoooks"
+import { useProjectTab } from "@/websocket-state/hooks/selectors/websocket-selector-hoooks"
 
 type SelectedFilesListProps = {
   selectedFiles: string[]
