+++ packages/client/src/components/tab-managers/projects-tab-manager.tsx
@@ -5,14 +5,14 @@ import {
   useUpdateProjectTab,
   useDeleteProjectTab,
   useUpdateSettings,
-} from '@/components/global-state/global-helper-hooks'
+} from '@/websocket-state/hooks/updaters/websocket-updater-hooks'
 import { GenericTabManager } from './generic-tab-manager'
 import { ShortcutDisplay } from '../app-shortcut-display'
 import { Input } from '../ui/input'
 import { Button } from '../ui/button'
 import { Pencil, Trash2 } from 'lucide-react'
 import { type ReactNode } from 'react'
-import { useActiveProjectTab, useAllProjectTabs, useSettings } from '../global-state/websocket-selector-hoooks'
+import { useActiveProjectTab, useAllProjectTabs, useSettings } from '@/websocket-state/hooks/selectors/websocket-selector-hoooks'
 
 type DialogContentProps = {
   tabId: string;
