+++ packages/client/src/components/tickets/ticket-dialog.tsx
@@ -12,7 +12,7 @@ import { TicketTasksPanel } from "./ticket-tasks-panel";
 import { useGetProjectFiles } from "@/hooks/api/use-projects-api";
 import { Checkbox } from "../ui/checkbox";
 import { ScrollArea } from "../ui/scroll-area";
-import { useCreateProjectTab } from "@/components/global-state/global-helper-hooks";
+import { useCreateProjectTab } from "@/websocket-state/hooks/updaters/websocket-updater-hooks";
 
 interface TicketDialogProps {
     isOpen: boolean;
