+++ packages/client/src/routes/file-search-interface.tsx
@@ -1,4 +1,4 @@
-import { useActiveProjectTab } from '@/components/global-state/websocket-selector-hoooks';
+import { useActiveProjectTab } from '@/websocket-state/hooks/selectors/websocket-selector-hoooks';
 import { useGeminiFileSearch } from '@/hooks/api/use-gemini-file-search';
 import { createFileRoute } from '@tanstack/react-router'
 import { useState } from 'react';
