+++ packages/client/src/websocket-state/hooks/selectors/use-global-state.tsx
